Create CVE-2020-26919.yaml
it was found that every section of the web could be used as a valid endpoint to submit POST requests being the action defined by the submitId argument. The problem was located in the login.html webpage, that has to be publicly available to perform login requests but does not implement any restriction for executing debug actions. This will allow users execute system commands.

Signed-off-by: GwanYeong Kim <gy741.kim@gmail.com>