Update sonarqube-login.yaml
The original template was causing too many false positives.  It seems that it was only picking off the BaseURL endpoint.  Meaning if any other service like Fortinet had a similar endpoint then it would show up as a Sonarqube result.

I've added matchers-condition as well as the HTTP status to 200 which eliminated the false positives in my case during testing.