fix(CVE-2024-55591): prevent false positive matches
The file tested by this template is served without the need of an
authentication token by some Fortinet services that are not FortiOS
Management interfaces. Those services are incorrectly reported as
vulnerable by this template.

This patch uses the same check than watchTowr Labs in their
[PoC](https://github.com/watchtowrlabs/fortios-auth-bypass-poc-CVE-2024-55591/blob/e1a468ece5636c6c10d1717b874773ef72fe3069/CVE-2024-55591-PoC.py#L136-L142)
to prevent false positive matches.